+++ src/Products/CMFCore/locales/update.sh
@@ -3,10 +3,10 @@
 # ``export PATH=$PATH:$BUILDOUT_DIR/bin`` when i18ndude is located in your buildout's bin directory)
 #
 # For every language you want to translate into you need a
-# locales/[language]/LC_MESSAGES/cmf_default.po
-# (e.g. locales/de/LC_MESSAGES/cmf_default.po)
+# locales/[language]/LC_MESSAGES/cmf_core.po
+# (e.g. locales/de/LC_MESSAGES/cmf_core.po)
 
-domain=cmf_default
+domain=cmf_core
 
 i18ndude rebuild-pot --pot $domain.pot --create $domain ../
 i18ndude sync --pot $domain.pot */LC_MESSAGES/$domain.po